+++ docs/spec/src/account/index.ts
@@ -1,0 +1,23 @@
+import { StepSpec } from '@jupiterone/integration-sdk-core';
+import { IntegrationConfig } from '../../../../src/config';
+
+export const accountSpec: StepSpec<IntegrationConfig>[] = [
+  {
+    /**
+     * ENDPOINT: n/a
+     * PATTERN: Singleton
+     */
+    id: 'fetch-account',
+    name: 'Fetch Account',
+    entities: [
+      {
+        resourceName: 'Account',
+        _class: ['Account'],
+        _type: 'xmcyber_account',
+      },
+    ],
+    relationships: [],
+    dependsOn: [],
+    implemented: true,
+  },
+];