+++ docs/spec/src/index.ts
@@ -1,8 +1,9 @@
 import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
+import { accountSpec } from './account';
 import { entitySpec } from './entity';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [...entitySpec],
+  integrationSteps: [...accountSpec, ...entitySpec],
 };