+++ src/steps/account/index.ts
@@ -1,0 +1,33 @@
+import {
+  IntegrationStepExecutionContext,
+  Step,
+} from '@jupiterone/integration-sdk-core';
+import { IntegrationConfig } from '../../config';
+import { Steps, Entities } from '../constants';
+import { createAccountEntity } from './converters';
+import { getAccount } from '../../initializeContext';
+
+export const ACCOUNT_ENTITY_KEY = 'entity:account';
+
+export async function fetchAccount(
+  context: IntegrationStepExecutionContext<IntegrationConfig>,
+): Promise<void> {
+  const { jobState } = context;
+  const accountEntity = await jobState.addEntity(
+    createAccountEntity(getAccount(context)),
+  );
+  await jobState.setData(ACCOUNT_ENTITY_KEY, accountEntity);
+}
+
+export const accountSteps: Step<
+  IntegrationStepExecutionContext<IntegrationConfig>
+>[] = [
+  {
+    id: Steps.ACCOUNT,
+    name: 'Fetch Account',
+    entities: [Entities.ACCOUNT],
+    relationships: [],
+    dependsOn: [],
+    executionHandler: fetchAccount,
+  },
+];