+++ src/steps/index.ts
@@ -1,5 +1,6 @@
 import { entitiesSteps } from './entities';
+import { accountSteps } from './account';
 
-const integrationSteps = [...entitiesSteps];
+const integrationSteps = [...accountSteps, ...entitiesSteps];
 
 export { integrationSteps };